+++ dirty_cat/minhash_encoder.py
@@ -7,7 +7,7 @@ The principle is as follows:
      n-dimensional vectors of integers.
   3. A hashing function is used to assign an integer to each n-gram.
      The minimum of the hashes over all n-grams is used in the encoding.
-  4. This process is repeated with N hashing functions are used to 
+  4. This process is repeated with N hashing functions are used to
      form N-dimensional encodings.
 Maxhash encodings can be computed similarly by taking the hashes maximum
 instead.
@@ -15,16 +15,10 @@ With this procedure, strings that share many n-grams have greater
 probability of having same encoding values. These encodings thus capture
 morphological similarities between strings.
 """
-import warnings
 
 import numpy as np
-from joblib import Parallel, delayed
-from scipy import sparse
-from sklearn.cluster import KMeans
-from sklearn.feature_extraction.text import CountVectorizer, HashingVectorizer
-from sklearn.neighbors import NearestNeighbors
 from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.utils import check_random_state, murmurhash3_32
+from sklearn.utils import murmurhash3_32
 
 from .fast_hash import ngram_min_hash
 from .utils import LRUDict, check_input
@@ -181,42 +175,47 @@ class MinHashEncoder(BaseEstimator, TransformerMixin):
             template = ("handle_missing should be either 'error' or "
                         "'zero_impute', got %s")
             raise ValueError(template % self.handle_missing)
-        X_out = np.zeros((len(X), self.n_components))
-        X = X.reshape(-1)
 
         # TODO Parallel run here
-        nan_idx = []
+        is_nan_idx = False
 
         if self.hashing == 'fast':
-            for i, x in enumerate(X):
-                if isinstance(x, float): # true if x is a missing value
-                    nan_idx.append(i)
-                elif x not in self.hash_dict:
-                    X_out[i, :] = self.hash_dict[x] = self.get_fast_hash(x)
-                else:
-                    X_out[i, :] = self.hash_dict[x]
-
+            X_out = np.zeros((len(X[:]), self.n_components * X.shape[1]))
+            counter = self.n_components
+            for k in range(X.shape[1]):
+                X_in = X[:, k].reshape(-1)
+                for i, x in enumerate(X_in):
+                    if isinstance(x, float): # true if x is a missing value
+                        is_nan_idx = True
+                    elif x not in self.hash_dict:
+                        X_out[i, k*self.n_components:counter] = self.hash_dict[x] = self.get_fast_hash(x)
+                    else:
+                        X_out[i, k*self.n_components:counter] = self.hash_dict[x]
+                counter += self.n_components
         elif self.hashing == 'murmur':
-            for i, x in enumerate(X):
-                if isinstance(x, float):
-                    nan_idx.append(i)
-                elif x not in self.hash_dict:
-                    X_out[i, :] = self.hash_dict[x] = self.minhash(
-                        x,
-                        n_components=self.n_components,
-                        ngram_range=self.ngram_range
-                    )
-                else:
-                    X_out[i, :] = self.hash_dict[x]
-
+            X_out = np.zeros((len(X[:]), self.n_components * X.shape[1]))
+            counter = self.n_components
+            for k in range(X.shape[1]):
+                X_in = X[:, k].reshape(-1)
+                for i, x in enumerate(X_in):
+                    if isinstance(x, float):
+                        is_nan_idx = True
+                    elif x not in self.hash_dict:
+                        X_out[i, k*self.n_components:counter] = self.hash_dict[x] = self.minhash(
+                            x,
+                            n_components=self.n_components,
+                            ngram_range=self.ngram_range
+                        )
+                    else:
+                        X_out[i, k*self.n_components:counter] = self.hash_dict[x]
+                counter += self.n_components
         else:
             raise ValueError("hashing function must be 'fast' or"
                              "'murmur', got '{}'"
                              "".format(self.hashing))
 
-        if self.handle_missing == 'error' and nan_idx:
+        if self.handle_missing == 'error' and is_nan_idx:
             msg = ("Found missing values in input data; set "
-                   "handle_missing='' to encode with missing values")
+                   "handle_missing='zero_impute' to encode with missing values")
             raise ValueError(msg)
-
         return X_out