+++ dirty_cat/similarity_encoder.py
@@ -208,7 +208,7 @@ class SimilarityEncoder(OneHotEncoder):
         Whether to raise an error or ignore if a unknown categorical feature is
         present during transform (default is to ignore). When this parameter
         is set to 'ignore' and an unknown category is encountered during
-        transform, the resulting one-hot encoded columns for this feature
+        transform, the resulting encoded columns for this feature
         will be all zeros. In the inverse transform, an unknown category
         will be denoted as None.
     handle_missing : 'error' or '' (default)
@@ -235,6 +235,8 @@ class SimilarityEncoder(OneHotEncoder):
     categories_ : list of arrays
         The categories of each feature determined during fitting
         (in order corresponding with output of ``transform``).
+    _infrequent_enabled: bool, default=False
+        Avoid taking into account the existance of infrequent categories.
 
     References
     ----------
@@ -404,6 +406,8 @@ class SimilarityEncoder(OneHotEncoder):
 
         if LooseVersion(sklearn.__version__) >= LooseVersion('0.21'):
             self.drop_idx_ = self._compute_drop_idx()
+        if LooseVersion(sklearn.__version__) >= LooseVersion('1.1.0'):
+            self._infrequent_enabled = False
 
         return self
 
@@ -540,3 +544,31 @@ class SimilarityEncoder(OneHotEncoder):
             out_row[:] = se_dict[x]
 
         return np.nan_to_num(out, copy=False)
+
+    def fit_transform(self, X, y=None, **fit_params):
+            """
+            Fit SimilarityEncoder to data, then transform it.
+            Fits transformer to `X` and `y` with optional parameters
+            `fit_params` and returns a transformed version of `X`.
+
+            Parameters
+            ----------
+            X : array-like of shape (n_samples, n_features)
+                Input samples.
+            y :  array-like of shape (n_samples,) or (n_samples, n_outputs), \
+                    default=None
+                Target values (None for unsupervised transformations).
+            **fit_params : dict
+                Additional fit parameters.
+
+            Returns
+            -------
+            X_new : ndarray array of shape (n_samples, n_features_new)
+                Transformed array.
+            """
+            if y is None:
+                # fit method of arity 1 (unsupervised transformation)
+                return self.fit(X, **fit_params).transform(X)
+            else:
+                # fit method of arity 2 (supervised transformation)
+                return self.fit(X, y, **fit_params).transform(X)